Remise au propre de l'étape 5 Respect du snake_case pour les commandes SQL Respect du camelCase pour le java Adaptation du package Service avec les DAO JDBC

+++ src/main/java/com/project/artconnect/dao/CommunityMemberDao.java
@@ -1,11 +1,12 @@
 package com.project.artconnect.dao;
 
 import com.project.artconnect.model.CommunityMember;
+
 import java.util.List;
 import java.util.Optional;
 
 public interface CommunityMemberDao {
-    Optional<CommunityMember> findById(Long id);
+    Optional<CommunityMember> findById(Integer id);
 
     List<CommunityMember> findAll();
 }
