+++ src/main/java/com/project/artconnect/dao/GalleryDao.java
@@ -1,11 +1,12 @@
 package com.project.artconnect.dao;
 
 import com.project.artconnect.model.Gallery;
+
 import java.util.List;
 import java.util.Optional;
 
 public interface GalleryDao {
-    Optional<Gallery> findById(Long id);
+    Optional<Gallery> findById(int id);
 
     List<Gallery> findAll();
 }
